fix: stop two plants from having overlapping interaction radiis due to one being activated
The fix using a scenario:
User1 "creates" Plant1 at (X1,Y1) but does not activate it. User2 later "creates" Plant2 at (X1,Y1) and activates it immediately.
Before this fix if User1 later tried to activate Plant1 the system would allow it. This is not good for the system.
This fix makes it so that when User1 later tries to activate Plant1, the operation fails. This is good for the system.