refactor: remove many unecessary occurences of the word 'all' in source code; fix: make useGetUserNearbyPlants and useGetUserPlants enabled even when lat and lon are falsy
Uncessary occurences of "all" was source code related so in places like function names and variables names. These changes were across the backend and frontend. For the plant query methods, lat and lon of zero are now used as defaults. This is okay because the frontend would hide any distance values and knows it is wrong